Java 13 compiler _runtime_ compatibility
This means that we're still compiling to Java 8, but the compiler itself
may be Java 13. This also means that the runtime during testing is Java
13.

A few adjustments needed to be made:
- Jacoco only supports Java 13 from v 0.8.5 upward
- Findbugs was updated to Spotbugs, which still needed a bump of asm lib
to v7.2. The _next_ release will depend on asm 7.2. There were still two
findings which we needed to ignore.
- Log4j1.2 has no support for the new java version strings and then
drops support for MDS because it thinks that we're running Java 1 :) We
updated to log4j2 for tests.