Make live test more reliable
This should actually be running with 128MiB of memory.
It was running with 96MiB, which was juuuuuust enough _if the timing was
right_. 128MiB will still support 500 clients which is a good tradeoff.

Weirdly, unless explicitly exposing 8080 on the testcontainer (it was
already declared in the Dockerfile), the library wouldn't even do a
health check at all, which could start the test before the container was
even started.